refactor: project page 구조 세분화 및 react-query를 사용한 서버 상태 관리
1. pages내 폴더 구조를 세분화
  - pages내 project 폴더를 생성하여, project와 관련된 페이지 파일
    접근성과 각 project 파일에 대한 인식률 향상을 도모
2. 내 project를 요청하는 함수를 react-query를 사용한 커스텀 훅으로 대체
  - react-query를 사용한 커스텀 훅으로 대체하여, 서버 상태 관리를 더욱
    용이하도록 만듦
