+++ FE/src/App.tsx
@@ -1,11 +1,11 @@
 import { Routes, Route } from 'react-router-dom';
-import ProjectPage from './pages/ProjectPage';
+import ProjectPage from './pages/project/ProjectPage';
 import LoginPage from './pages/LoginPage';
 import MainPage from './pages/MainPage';
 import BacklogPage from './pages/BacklogPage';
 import SprintPage from './pages/SprintPage';
 import ReviewPage from './pages/ReviewPage';
-import ProjectCreatePage from './pages/ProjectCreatePage';
+import ProjectCreatePage from './pages/project/ProjectCreatePage';
 import { ModalProvider } from './modal/ModalProvider';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
