+++ FE/src/apis/Project/index.ts
@@ -1,0 +1,1 @@
+export { default as fetchGetMyProjects } from './fetchGetMyProjects';
