+++ FE/src/constants/constants.tsx
@@ -5,3 +5,7 @@ export const REFRESH_TOKEN = 'refreshToken';
 export const PROCESS1 = 0;
 
 export const PROCESS2 = 1;
+
+export const API_URL = {
+  PROJECT: '/projects',
+};
