+++ FE/src/hooks/queries/project/index.ts
@@ -1,0 +1,1 @@
+export { default as useGetMyProjects } from './useGetMyProjects';
