+++ FE/src/hooks/queries/project/useGetMyProjects.tsx
@@ -1,0 +1,12 @@
+import { useQuery } from '@tanstack/react-query';
+import { fetchGetMyProjects } from '../../../apis/Project';
+
+const useGetMyProjects = () =>
+  useQuery({
+    queryKey: ['myProjects'],
+    queryFn: fetchGetMyProjects,
+    refetchOnMount: false,
+    refetchOnWindowFocus: false,
+  });
+
+export default useGetMyProjects;
