+++ FE/src/pages/project/ProjectCreatePage.tsx
@@ -1,8 +1,8 @@
 import { Link } from 'react-router-dom';
-import chevronLeft from '../assets/icons/chevron-left.svg';
-import ProjectProcessText from '../components/project/ProjectProcessText';
-import ProjectCreateForm from '../components/project/ProjectCreateForm';
-import useProjectForm from '../hooks/pages/project/useProjectForm';
+import chevronLeft from '../../assets/icons/chevron-left.svg';
+import ProjectProcessText from '../../components/project/ProjectProcessText';
+import ProjectCreateForm from '../../components/project/ProjectCreateForm';
+import useProjectForm from '../../hooks/pages/project/useProjectForm';
 
 const ProjectCreatePage = () => {
   const { process, handleNextButtonClick, handleSubmitButtonClick, projectNameRef, projectSubjectRef } =
